Add no-assert "smoke tests" for UI functions that don't return a UIComm
In the absence of a UIComm object, there's not really much to test (on
the UI side anyway). However, the tests are still valuable as a "light
integration test" for the backend modules they call: by exercising some
of their code paths, they at least require that the backend modules
don't crash in these simple cases.

Clearly, more serious testing of the end-to-end functionality of these
commands will need to be done in the real integration tests. These are
just "smoke tests" of a sort.